requirements.txt
PyMuPDF>=1.23.0
Pillow>=9.0.0
openai>=1.0.0
python-dotenv>=1.0.0
psutil>=7.0.0